Update Apps
I guess there is a new field consequently while listing all apps validation fails. I don't know what this filed is for and couldn't found it in the documentation. Maybe documentation is not updated to reflect the changes yet. Because of this left is as Optional. Thanks for the awesome code by the way